Removed double tap event and handling preview inside the adapter
